Se agrego endpoint para poder hacer reservas a nombre de un cliente tanto como empresa y gestor
Se modifico la relacion entre las tablas perfil empresa y perfil gestor para asi mismo poder asignar mas gestores a una empresa

Se modifico el controlador de reservas para que el costoCancha y costoTotal no se pase por body

Se agrego endpoint nuevo para poder reservar dos canchas en un solo proceso

+++ Back_End/src/main/java/com/example/sport_full/models/ReservationsModels.java
@@ -43,7 +43,7 @@ public class ReservationsModels {
         CANCELADA
     }
 
-    private Long costoHora;
+    private Double costoHora;
 
     private Long costoTotal;
 
@@ -96,11 +96,11 @@ public class ReservationsModels {
         this.fechaPago = fechaPago;
     }
 
-    public Long getCostoHora() {
+    public Double getCostoHora() {
         return costoHora;
     }
 
-    public void setCostoHora(Long costoHora) {
+    public void setCostoHora(Double costoHora) {
         this.costoHora = costoHora;
     }
